+++ doordeck-sdk/src/jvmMain/kotlin/com/doordeck/multiplatform/sdk/model/data/FusionOperations.jvm.kt
@@ -65,7 +65,7 @@ object FusionOperations {
     ) : LockController
 
     data class DemoController(
-        val port: Int = 8080
+        val port: UShort = 8080u
     ) : LockController
     
     data class GallagherController(
